feat: compress objects while writing to S3, uncompress while reading
This makes use of the async-compression crate. Rather than drag in all
the algorithms it supports, we initially use two, zstd and bzip2, but
it's trivial to add support for the other async-compression
algorithms, e.g. gzip, brotli, deflate, lzma, xz, zlib, etc.

Any SBOM written via the api with a Content-Type of `application/json`
will be compressed with the zstd encoder while being streamed to
S3. Any S3 object with a Content-Type of `application/zstd` fetched
via the api will be uncompressed with the zstd decoder as it is
streamed back to the api client.

Further, any SBOM's with a Content-Type *other* than
`application/json` will be stored as is, with the assumption that
they're already encoded/compressed, e.g. bzip2. When fetched through
the api, this content will be uncompressed with the appropiate decoder
to `application/json` as it's streamed back to the client.

Signed-off-by: Jim Crossley <jim@crossleys.org>