fix: unify handling of storage events from minio and s3
The s3 and minio paths diverged as they were relying on separate formats. Looking more detailed at the formats showed that other parts of the minio data could be relied on which were encoded the same way as s3. This way there is no need to distinguish between MinIO and S3 storage events.