fix: use content-type and content-encoding correctly
We need to be able to decode whatever encoded payload we might accept,
so the previous logic of accepting *any* content-type uploaded was
flawed. For now, we require json and we always encode it as zstd while
writing to S3.

Once Compress middleware is installed, we'll change the storage API to
prevent needless decoding from S3 prior to streaming back to the HTTP
client.

Signed-off-by: Jim Crossley <jim@crossleys.org>