feat: use multivalues instead of multiple documents
This changes the document model to one that treats 1 SBOM as 1 document, and 1 VEX as 1 document. Matching inside dependencies and CVEs will still happen as before, but
the information returned from search now corresponds to a particular SBOM or VEX document. This creates a consistency in the UI as discussed in #150