fix: metadata can now be attached to large, multipart SBOM's
Fixes #161

This required a fix in the rust-s3 crate so we're now depending on a
branch of a fork of that repo.

The fork's API differs slightly from the version we were previously
using: we can now report the size of the upload instead of just the
status code.

And while I was at it, I removed a confusing debug println that should
clean up our server log a bit.

Signed-off-by: Jim Crossley <jim@crossleys.org>