feat: avoid port conflicts when running concurrent integs
Introduce a `serve` function that creates and binds the actix http
server to a port but doesn't await it. This enables a client to obtain
the OS-assigned port in the `Run` struct before starting the server.

Note on rebasing after the metrics changes: passing Run struct fields
to configure/serve fn's feels wrong. Ideally, I think I'd like for
infra.run() to return a Server I can query instead of an opaque
future.

Signed-off-by: Jim Crossley <jim@crossleys.org>