fix: differentiate between backend and frontend URLs
The realm can be configured to use a different URL for the frontend.
This is important when running inside containers, where (between
services) hostnames are different than for the end-user. The problem
is with the "issuer URL" which get's acquired through the frontend
path, but is then validated through the backend path. URLs don't match
and so the access token is rejected.

This change allows to configure the frontend URL and makes of this for
the local compose deployment.