fix: add Z option to volume mount as an env var
This commit adds the Z option to the volume mount for init-sso.

The motivation for this change is that on Linux systems the SELinux
context for the volume mount is not set correctly and init-keycloak
container is not able access the volume mount.
This can been seen by the following erorr in the logs:
```console
[minio]
podman start -a compose_init-keycloak_1
/usr/bin/bash: /init-sso/init.sh: Permission denied
```

Adding this option as an environment variable should fix this issue
and allows the init-keycloak container to access the volume mount.

Refs: https://github.com/danbev/learning-crypto/blob/main/notes/seedwing/podman-compose-issue.md

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>