doc: update instructions for starting containers
This commit updates the instructions for starting the containers locally
and specifically that the keycloak container needs to be started first
and fully initialized before the other containers are started.

The motivation for this change is that the containers that depend on
keyclaok will fail because the
http://keycloak:8080/realms/chicken/.well-known/openid-configuration
endpoint will not be available that the time they start.

At least this is the behavior that I seeing on Linux and I think it
would help others starting up the project to have this information in
the README as a possible workaround.

Refs: https://github.com/danbev/learning-crypto/blob/main/notes/seedwing/podman-compose-issue.md#second-issue
Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>