fix: add wait for keycloak in compose-trustification
This commit adds a wait for keycloak to be up and running before
starting the trustification services. This is done by adding a
command to the services that will wait for keycloak to be up and
running before starting the trustification services.

The motivation for this is that currently if one tries run the following
command:
```console
$ podman-compose -f compose.yaml -f compose-trustification.yaml up
```
This would lead to the trustification services failing to start.

With this commit and using the same command as above we should see
something like the following while it is running:
```console
$ podman logs -f compose_spog-api_1
waiting for keycloak...
waiting for keycloak...
waiting for keycloak...
keycloak is up
[2023-08-22T06:52:17.329Z INFO  trustification_infrastructure::infra] Setting up infrastructure endpoint
[2023-08-22T06:52:17.330Z INFO  trustification_infrastructure::infra] Running infrastructure endpoint on:
[2023-08-22T06:52:17.330Z INFO  trustification_infrastructure::infra]    http://[::1]:9010
[2023-08-22T06:52:17.330Z INFO  trustification_infrastructure::infra]    http://127.0.0.1:9010
```

This somewhat messy at the moment but if this works we can look at
cleaning it up.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>