feat: point the integs at a remote trustification server
Fixes #401

Setting the TRUST_URL env var to a remote trustification server
triggers the integs to be run against it.

If it's set, other env vars will be required:
  TRUST_USER_ID -- the client id of the user
  TRUST_MANAGER_ID -- the client id of the manager
  TRUST_SECRET -- the secret is assumed to be the same for user & mgr

If KAFKA_BOOTSTRAP_SERVERS is set, its value will be used to configure
the event bus. Otherwise, SQS is assumed and valid AWS credentials
will be required.

Signed-off-by: Jim Crossley <jim@crossleys.org>