fix: add Self::Provided match arm props_query
This commit adds a match arm for the SearchPropertiesMode::Provided
variant in SearchPropertiesMode::props_query.

The motivation for this is that when an AdvisorySearch component is
instantiated it will be passed a SearchPropertiesMode::Provided variant:
```rust
<AdvisorySearch callback={advisory_callback}
     mode={SearchPropertiesMode::Provided {terms: (*search_terms).clone()}}>
    <AdvisoryResult state={(*advisory_search).clone()} />
</AdvisorySearch>
```
Notice here that the `mode` property of the `AdvisorySearch` component
is set to `SearchPropertiesMode::Provided` and the term is set to the
search term. But if we look at the current implementation of the
SearchPropertiesMode::props_query method we see that it only returns
the query if the mode is Managed:
```rust
pub enum SearchPropertiesMode {
    Managed { query: Option<String> },
    Provided { terms: String },
}

impl SearchPropertiesMode {
    pub fn props_query(&self) -> Option<String> {
        match &self {
            Self::Managed { query } => query.clone(),
            _ => None,
        }
    }
}
```
When props_query is later called by AdvisorySearch::advisory_search
and SbomSearch::sbom_search, the query term will not be available as
props_query will return None, and the search will be performed without
a query parameter which will return all the advisories and sboms.

Fixes: https://github.com/trustification/trustification/issues/386
Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>