fix: don't show all advisories in case there's no identifier
In case CPE and PURL are missing, it will currently result in an
unbounded search, returning "everything". This is misleading and not
useful.

This change makes the number of entries optional, preventing the search
in the first place, and also hiding the number altogether, as "zero"
would be misleading as well.