fix: retry objects that fail fetching to avoid failing retrieval process early
When walking a large number of objects, existing early because of a temporary network error
has some major implications on the rest of the system, such as needing to restart the process of walking the objects
or keeping tracked of the failed ones.

Instead, retry objects that failed to fetch a number of times before aborting.