chore: merge trustification.dev repository into trustification
The primary reason is that the automation often looks at git commits and use images with that sha, but will also use that when propagating changes from staging -> prod. So by keeping these separate repos we create ourselves more friction than necessary. It felt nice to keep the helm charts separate in the beginning, but I've also seen a few issues when it comes to keeping code in sync with the charts, so there is a benefit from keeping it together.