feat: propagate s3 events to indexed queue
Slightly change the semantics of the indexed queue so that it includes the full event metadata, and rather than indicating a successful indexing, it contains the full event. This allows chaining other processes on the indexed queue that
want to react to events that are processed by the indexer (even if failed).