fix: don't let the default override any use made changes on "back"
So far, we did store the page state in the browser history. However,
when unpacking it (on hitting "back"), the defaults did override any
selection made by the user so far.

This change fixes this by tracking both the "modification" state as well
as the "defaults" which need to get applied. Also when the user
resets the search.