fix: always reindex bombastic at startup
Deleting documents from bombastic will not delete entries from the packages index,
because packages are unique and index does not keep relationship to the SBOM (since multiple SBOMs may contain the same package).

As a workaround, reindexing the bombastic index on startup will allow the package
index to be rebuilt upon a restart from the SBOMs that exist in the archive.