[stable] Cherry-pick relevant commits of PR #1038 (#1039)
* chore: clear up logging a bit

* chore: align naming

* refactor: extract handling re-index into its own function

* fix: prevent empty indexes when using --reindex always

When reindexing, the indexes are "reset", which basically means that
new indexes are being created. However, the loop still holds on to the
old writers. So when resetting the indexes, we must also create new
writers.

Otherwise the logic will write to the old indexes, and the new ones
will stay empty.