fix: ensure that the storage key does not alter the path structure
The key has a direct impact on the path inside the S3 bucket. We must
ensure that the user is not able to influence the location of the
file, only the name of it.

This change encodes the provided key using urlencoding, including the
slash. This prevents the user providing a file-name which can escape
the path structure.