[release/1.0.z] backport: #1269 #1270 #1272 (#1274)
* fix: provide a way to retrieve public vs non-public frontend config

Some part of the configuration might be considered confidential. So
we do remove most of the configuration information and create a
dedicated, public endpoint.

* docs: document the use of a custom trust anchor

* fix: ensure that the storage key does not alter the path structure

The key has a direct impact on the path inside the S3 bucket. We must
ensure that the user is not able to influence the location of the
file, only the name of it.

This change encodes the provided key using urlencoding, including the
slash. This prevents the user providing a file-name which can escape
the path structure.