Allow list of certs to be separated by comma
Allow list of TLS certificates to be separated with comma. With
this change, it can be possible to specify a list of TLS certs
in `CLIENT_TLS_CA_CERTIFICATES` as a CSV list. Prior this change,
`CLIENT_TLS_CA_CERTIFICATES` can handle only one certificate and
the only way to specify a list of certificates was by using a
consecutive sequence of `--client-tls-ca-certificates` options.

Signed-off-by: Jiří Kučera <jkucera@redhat.com>