Increase a livenessProbe delay
The short initial delay is cause a problem for pods that require an
access to the index. The 2 seconds isn't enough for bombastic-api to
properly start. Increasing it to 30 should give an app enough time.

Having a low values here causes the K8s kills a pod before it can start
properly.

JIRA: ISV-1880

Signed-off-by: Ales Raszka <araszka@redhat.com>